requirements.txt
Pillow==8.2.0
pytest==6.2.3
